feat: add initial backend setup with user auth models, repos, and security config
- Added ApplicationUser and Role entities
- Implemented RoleRepository and UserRepository
- Created UserService and AuthenticationController
- Configured basic Spring Security (permit all for now)
- Added application bootstrap with initial user and role seeding

+++ xitter-backend/src/main/java/com/xitter/models/Role.java
@@ -13,7 +13,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "role_id")
-    private Long roleId;
+    private Integer roleId;
 
     private String authority;
 
@@ -21,17 +21,17 @@ public class Role {
         super();
     }
 
-    public Role(Long roleId, String authority) {
+    public Role(Integer roleId, String authority) {
         super();
         this.roleId = roleId;
         this.authority = authority;
     }
 
-    public Long getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Long roleId) {
+    public void setRoleId(Integer roleId) {
         this.roleId = roleId;
     }
